pyproject.toml
[tool.poetry]
name = "covidbot"
version = "0.1.0"
description = ""
authors = ["sovitpy <sovitlekhak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
tweepy = "3.8.0"
beautifulsoup4 = "4.9.1"
schedule = "0.6.0"
discord-webhook = "^0.16.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
